requirements.txt
acme==2.11.0
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.1
celery==5.4.0
certbot==2.11.0
certbot-nginx==2.11.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
ConfigArgParse==1.7
configobj==5.0.8
cryptography==43.0.1
distro==1.9.0
dj-rest-auth==6.0.0
Django==5.1.2
django-allauth==65.0.2
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.8
importlib_metadata==8.5.0
importlib_resources==6.4.5
josepy==1.14.0
kombu==5.4.2
packaging==24.1
parsedatetime==2.6
paypal-checkout-serversdk==1.0.3
paypalhttp==1.0.1
paypalrestsdk==1.13.3
prompt_toolkit==3.0.47
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
pyRFC3339==1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
requests==2.32.3
setuptools==75.2.0
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.2
vine==5.1.0
vue==0.0.1
wcwidth==0.2.13
zipp==3.20.1
